Renames + some login strategy service test updates
