+++ libs/auth/src/common/login-strategies/sso-login.strategy.spec.ts
@@ -39,7 +39,7 @@ import {
 } from "../abstractions";
 import { SsoLoginCredentials } from "../models/domain/login-credentials";
 
-import { identityTokenResponseFactory } from "./login.strategy.spec";
+import { identityTokenResponseFactory } from "./base-login.strategy.spec";
 import { SsoLoginStrategy } from "./sso-login.strategy";
 
 describe("SsoLoginStrategy", () => {
